Auto-patching build part VI: Auto convert all code from K&R to ANSI using cproto.  Doing this required fixing up function signature mismatches throughout the demos.

+++ arena/main.c
@@ -11,7 +11,7 @@
 #define EM_CHILD_APP_NAME EM_ARENA
 #include "EM_CHILD_APP.h"
 
-int screen_size[2];
+long screen_size[2];
 
 char buf[80];
 double zoom = 1.0;
